% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{arid_ind_tables}
\alias{arid_ind_tables}
\title{Aridity index}
\format{list formed by six data frames.}
\usage{
data(Trent_climate)
}
\description{
Used for reference in aridity indices assessment (see function \code{\link{arid}} and
references for data sources).
}
\keyword{datasets}

