% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{clima_81_10}
\alias{clima_81_10}
\title{Climatic normals of precipitation and temperatures}
\format{A list (one table for each station) of 28 monthly climatic normals}
\usage{
data(Trent_climate)
}
\description{
Climatic normals of precipitation and temperature (minimum, maximum, and mean) for the 
climatic period 1981 - 2010. It has been calculated by function \code{\link{climate}}.
}
\keyword{datasets}

