% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{coeff_rad}
\alias{coeff_rad}
\title{Radiative energy coefficients}
\format{
An array of 12 numerics
}
\usage{
data(Trent_climate)
}
\description{
"Radiative energy coefficients" for Hargreaves' equation, corresponding to the 
daily extra-atmospheric solar radiation energy. It is the output of function 
\code{\link{ExAtRa}}.
}
\keyword{datasets}
