% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{continental_ind_tables}
\alias{continental_ind_tables}
\title{continentality/oceanicity indices}
\format{list of 4 data frames}
\usage{
data(Trent_climate)
}
\description{
Used for reference in continentality/oceanicity indices assessment (see function
\code{\link{contin}} and references for data sources).
}
\keyword{datasets}

