% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{coord_elev}
\alias{coord_elev}
\title{Geographical position for each meteorological station}
\format{Data frame of 40 observations of 4 variables}
\usage{
data(Trent_climate)
}
\description{
Coordinates and elevation for each station in the data set. Coordinates are geographical 
and elevation is measured in meters above mean sea level.
}
\keyword{datasets}

