% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{P}
\alias{P}
\title{Precipitation}
\format{Data frame: 19358 obervations of 39 variables (stations)}
\usage{
data(Trent_climate)
}
\description{
The daily data frame of precipitation for a number of stations. It is used in function  
\code{\link{oiv_ind}}
}
\keyword{datasets}

