% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{quantiles}
\alias{quantiles}
\title{Monthly quantiles of the meteorological variables}
\format{
Data frame of 28 observations of 4 variables
}
\usage{
data(Trent_climate)
}
\description{
The second list (\code{quantiles}) in \code{thornt_lst} organized according to stations. 
See Examples in function \code{\link{thornthwaite}} for its construction.
}
\keyword{datasets}
