% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{thornt_lst}
\alias{thornt_lst}
\title{input for the Thornthwaite function}
\format{S3 object: a "hyperlist" (list of lists of lists), one list of lists for each station}
\usage{
data(Trent_climate)
}
\description{
For every station, the first element (a list, \code{Thornth._W._bal}) reports the monthly
series of water balance quantities for the station, each in one data frame (see function
\code{\link{thornthwaite}} for details). The second list (\code{quantiles}) reports the
monthly quantiles for the same quantities.
}
\keyword{datasets}

