% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{Tm}
\alias{Tm}
\title{Mean daily temperature}
\format{data frame: 19358 obervations of 15 variables (stations)}
\usage{
data(Trent_climate)
}
\description{
the daily data frame of mean daily temperature for a number of stations, used in function
 \code{\link{oiv_ind}}
}
\keyword{datasets}

