% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{Tn}
\alias{Tn}
\title{Minimum daily temperature}
\format{
Data frame: 19358 obervations of 15 variables (stations)
}
\usage{
data(Trent_climate)
}
\description{
The daily data frame of minimum daily temperature for a number of stations, used in function 
 \code{\link{oiv_ind}}
}
\keyword{datasets}
