% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{Tx}
\alias{Tx}
\title{Maximum daily temperature}
\format{Data frame: 19358 obervations of 15 variables (stations)}
\usage{
data(Trent_climate)
}
\description{
The daily data frame of maximum daily temperature for a number of stations, used in function 
 \code{\link{oiv_ind}}
}
\keyword{datasets}

