% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{W_balance}
\alias{W_balance}
\title{Water balance}
\format{list of 28 elements, each is a data frame of 5 variables and 636 observations}
\usage{
data(Trent_climate)
}
\description{
is the first list (\code{W_balance}) in \code{thornt_lst} organized according to stations.
See Examples in function \code{\link{thornthwaite}} for its construction.
}
\keyword{datasets}

