% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{coord_elev}
\alias{coord_elev}
\title{geographical position for each meteorological station
coordinates and elevation for each station in the data set. Coordinates are geographical
and elevation is measured in meters above mean sea level.}
\format{data frame of 40 observations of 4 variables}
\usage{
data(Trent_climate)
}
\description{
geographical position for each meteorological station
coordinates and elevation for each station in the data set. Coordinates are geographical
and elevation is measured in meters above mean sea level.
}
\keyword{datasets}

