% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{lista_cli}
\alias{lista_cli}
\title{Dataset of meteorological measures}
\format{list of 28 elements, each is a data frame of 5 variables and 636 observations}
\usage{
data(Trent_climate)
}
\description{
A list of 40 data frames (one for each station of the meteoroglical network), with
monthly time series of precipitation and temperature (minimum and maximum).
}
\keyword{datasets}

