% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{quantiles}
\alias{quantiles}
\title{Monthly quantiles of the meteorological varaibales}
\format{data frame of 40 observations of 4 variables}
\usage{
data(Trent_climate)
}
\description{
the second list (\code{quantiles}) in \code{thornt_lst} organized according to stations.
See Examples in function \code{\link{thornthwaite}} for its construction.
}
\keyword{datasets}

