% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Trent_climate.R
\docType{data}
\name{Trent_climate}
\alias{Trent_climate}
\title{Data set of Trentino climate}
\usage{
data(Trent_climate)
}
\description{
A group of datasets used consistently throughout the ClimClass manual and examples.
It is used as reference definition of the climate for the Trentino region, Italy.
It includes monthly series of temperature and precipitation, and reference tables for
the definition of aridity and continentality/oceanicity
}
\details{
Series like "Txxxx" were supplied by the Autonomous Province of Trento - Meteotrentino (I).
Series like "FEMxx" were supplied by Fondazione Edmund Mach, San Michele all'Adige (I).
}
\keyword{datasets}

